PythonServer/requirements.txt
# Python Backend Dependencies
# Updated for Python 3.13 compatibility

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0

# LangChain ecosystem
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.5.0

# Google AI
google-generativeai>=0.8.0

# LangGraph for Agent
langgraph>=0.2.0

# PDF Processing
pypdf>=4.0.0
